rename utils to cached_property
