+++ dephell/cached_property.py
