+++ dephell/controllers/_readme.py
@@ -10,7 +10,7 @@ from m2r import convert
 
 # app
 from ..constants import EXTENSIONS
-from ..utils import cached_property
+from ..cached_property import cached_property
 
 
 CODE = """
