+++ dephell/models/extra_dependency.py
@@ -3,7 +3,7 @@ import attr
 from packaging.utils import canonicalize_name
 
 # app
-from ..utils import cached_property
+from ..cached_property import cached_property
 from .dependency import Dependency
 from .groups import Groups
