+++ dephell/repositories/_git/bitbucket.py
@@ -6,7 +6,7 @@ import requests
 
 # app
 from ...config import config
-from ...utils import cached_property
+from ...cached_property import cached_property
 from .base import BaseRepo
