+++ dephell/repositories/_git/git.py
@@ -13,7 +13,7 @@ from ...config import config
 from ...context_tools import chdir
 from ...models.git_release import GitRelease
 from ...models.release import Release
-from ...utils import cached_property
+from ...cached_property import cached_property
 from ..base import Interface
 from .._local_single import LocalRepo
